refactor: extract fixation utilities
